Remove functions from the non-test build.
